bot/requirements.txt
python-telegram-bot[all]==21.5
sqlalchemy==2.0.23
python-dotenv==1.0.0
openai==1.3.5
pillow==10.1.0
qrcode==7.4.2
pyyaml==6.0.1
alembic==1.12.1
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
